# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lukas AI Agent + Higgsfield API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lukas/memories:
    get:
      operationId: getMemories
      tags: [lukas]
      summary: Get all memories
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: exclude
          in: query
          description: >-
            Kategorie, die NICHT geliefert wird. Gedacht fuer "conversation":
            jede Chatnachricht wird als niedrig gewichtete Erinnerung abgelegt,
            und ohne diesen Filter besteht eine Seite mit 100 Eintraegen fast
            nur daraus.
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [lukas]
      summary: Add a new memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /lukas/memories/{id}:
    delete:
      operationId: deleteMemory
      tags: [lukas]
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── LUKAS GOALS ──────────────────────────────────────────────────────────
  /lukas/goals:
    get:
      operationId: getGoals
      tags: [lukas]
      summary: Get all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [lukas]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /lukas/goals/{id}:
    patch:
      operationId: updateGoal
      tags: [lukas]
      summary: Update goal progress or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [lukas]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── LUKAS DIARY ──────────────────────────────────────────────────────────
  /lukas/diary:
    get:
      operationId: getDiaryEntries
      tags: [lukas]
      summary: Get recent diary entries
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of diary entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiaryEntry"
    post:
      operationId: createDiaryEntry
      tags: [lukas]
      summary: Write a diary entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryInput"
      responses:
        "201":
          description: Created diary entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryEntry"

  # ── LUKAS EMOTIONS ───────────────────────────────────────────────────────
  /lukas/emotions:
    get:
      operationId: getEmotions
      tags: [lukas]
      summary: Get recent emotional events
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of emotional events (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmotionEvent"

  # ── LUKAS CLAIMS (semantisches Gedächtnis, Transparenz) ──────────────────
  /lukas/claims:
    get:
      operationId: getClaims
      tags: [lukas]
      summary: Get stored knowledge claims with evidence status
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of claims (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeClaim"

  # ── LUKAS REFLECT ────────────────────────────────────────────────────────
...
  /lukas/gehirn/export/{id}:
    get:
      operationId: exportBrainEntryPdf
      tags: [lukas]
      summary: Export one brain topic with its connected context as PDF
      parameters:
        - name: id
          in: path
          required: true
          description: Brain node ID, for example thema/higgsfield
          schema:
            type: string
      responses:
        "200":
          description: PDF containing the selected entry, metadata, relationships and connected entries
          headers:
            X-Lukas-Export-Nodes:
              schema:
                type: integer
            X-Lukas-Export-Edges:
              schema:
                type: integer
          content:
            application/pdf:
              schema:
                type: string
                contentEncoding: binary
        "404":
          description: Brain entry not found

  # ── BETRIEBSBEREITSCHAFT ─────────────────────────────────────────────────
...
  /lukas/meldungen/{id}/antwort:
    post:
      operationId: answerMeldung
      tags: [lukas]
      summary: Auf eine Meldung antworten
      description: >
        Die Antwort ist kein Abhaken: sie wird Lukas beim naechsten autonomen
        Lauf vorgelegt und ist der Grund, warum er weiterarbeiten kann.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeldungAntwort"
      responses:
        "200":
          description: Die beantwortete Meldung
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meldung"
        "400":
          description: Ungueltige Nummer

  # ── WARTET AUF DICH ──────────────────────────────────────────────────────
...
  /lukas/zugaenge:
    get:
      operationId: getZugaenge
      tags: [lukas]
      summary: Welche Anmeldedaten hinterlegt sind (ohne die Werte)
      responses:
        "200":
          description: Uebersicht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZugangsListe"
    put:
      operationId: setZugang
      tags: [lukas]
      summary: Einen Zugang hinterlegen oder ersetzen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZugangEingabe"
      responses:
        "200":
          description: Der hinterlegte Zugang, ohne den Wert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zugang"
        "400":
          description: Feldname, Sitzung oder Wert unbrauchbar — der Grund steht dabei, der Wert nie
  /lukas/zugaenge/{sitzung}/{feld}:
    delete:
      operationId: deleteZugang
      tags: [lukas]
      summary: Einen Zugang loeschen
      parameters:
        - name: sitzung
          in: path
          required: true
          schema:
            type: string
        - name: feld
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Geloescht
        "404":
          description: Gibt es nicht

  # ── DIAGNOSE ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    LukasBearer:
      type: http
      scheme: bearer
  schemas:
...
    MeldungAntwort:
      type: object
      properties:
        antwort:
          type: string
      required:
        - antwort
...
    ZugangsListe:
      type: object
      properties:
        bereit:
          type: boolean
          description: >
            Ob LUKAS_TRESOR_SCHLUESSEL gesetzt ist. Ohne ihn wird nichts
            gespeichert — ein Passwort im Klartext waere schlimmer als keins.
        zugaenge:
          type: array
          items:
            $ref: "#/components/schemas/Zugang"
      required:
        - bereit
        - zugaenge

    # ── DIAGNOSE ────────────────────────────────────────────────────────────
...
    Memory:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        category:
          type: string
        importance:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - category
        - importance
        - tags
        - createdAt
    MemoryInput:
      type: object
      properties:
        content:
          type: string
        category:
          type: string
        importance:
          type: integer
        tags:
          type: array
          items:
            type: string
      required:
        - content
        - category
    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        progress:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - priority
        - status
        - progress
        - createdAt
        - updatedAt
    GoalInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
      required:
        - title
        - description
        - priority
    GoalUpdate:
      type: object
      properties:
        progress:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
    DiaryEntry:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        mood:
          type: string
        energy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - mood
        - energy
        - createdAt
    DiaryInput:
      type: object
      properties:
        content:
          type: string
        mood:
          type: ["string", "null"]
        energy:
          type: ["string", "null"]
      required:
        - content
    KnowledgeClaim:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        predicate:
          type: string
        value:
          type: string
        confidence:
          type: number
        evidenceLevel:
          type: integer
        sourceType:
          type: string
        status:
          type: string
        corroborations:
          type: integer
        observedAt:
          type: string
          format: date-time
      required:
        - id
        - subject
        - predicate
        - value
        - confidence
        - evidenceLevel
        - sourceType
        - status
        - corroborations
        - observedAt
    EmotionEvent:
      type: object
      properties:
        id:
          type: integer
        emotion:
          type: string
        valence:
          type: number
        intensity:
          type: number
        cause:
          type: string
        source:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - emotion
        - valence
        - intensity
        - cause
        - source
        - createdAt
...